Organized, Cleaned Up, Broke the code
I organized everything into separate folders, cleaned up names and functionality. However, by reorganizing everything into separate folders, I broke my own code. Is being fixed.
