+++ aec_velocity.h
@@ -1,0 +1,16 @@
+//
+//  aec_velocity.h
+//  arrental_engine
+//
+//  Created by Jared Rand on 2017-05-07.
+//  Copyright © 2017 zapix_games. All rights reserved.
+//
+
+#ifndef aec_velocity_h
+#define aec_velocity_h
+
+#include "ae_components.h"
+
+void AEC_Velocity_UpdateSimple(AEC_EntityCatalog* entityCatalog, unsigned int entity_at);
+
+#endif /* aec_velocity_h */
